Edit and BookMark methods [WIP]

+++ Controllers/BooksController.cs
@@ -11,7 +11,7 @@ using Microsoft.Extensions.FileSystemGlobbing.Internal.PatternContexts;
 namespace BearPawPages.Controllers
 {
     public class BooksController : Controller
-        
+
     {
         private BookDbContext context;
 
@@ -38,7 +38,7 @@ namespace BearPawPages.Controllers
         [HttpPost]  //PROCESSES form
         public IActionResult Add(AddBookViewModel addBookViewModel)
         {
-           if(ModelState.IsValid)
+            if (ModelState.IsValid)
             {
                 Book newBook = new Book
                 {
@@ -46,7 +46,7 @@ namespace BearPawPages.Controllers
                     Author = addBookViewModel.Author,
                     TotalPage = addBookViewModel.TotalPage,
                     CurrentPage = addBookViewModel.CurrentPage,
-                    //ReadingDate = addBookViewmodel.ReadingDate,
+                    ReadingDate = addBookViewModel.ReadingDate,
                     ReadingNotes = addBookViewModel.ReadingNotes
                 };
 
@@ -61,9 +61,10 @@ namespace BearPawPages.Controllers
 
         }
 
-        //Edit and EditBookData UPDATE all properties of book objects
+        //Edit and EditBookData UPDATE ALL properties of book objects
         //GET
         [Route("/Books/Edit/{bookId}")]
+
         public IActionResult Edit(int bookId)
         {
             Book book = context.Books.Find(bookId);
@@ -73,78 +74,85 @@ namespace BearPawPages.Controllers
             return View();
         }
 
+        //TODO: Work on this method.  Once this one is working, 
+        //apply it to the BookMark method
         [HttpPost]
         [Route("/Books/Edit")]
-        public IActionResult EditBookData(int bookId, string title, string author, int totalPage, int currentPage, DateTime readingDate, string readingNotes)
+        public IActionResult EditBookData(EditBookViewModel editBookViewModel)
         {
-            Book book = context.Books.Find(bookId);
-            book.Title = title;
-            book.Author = author;
-            book.TotalPage = totalPage;
-            book.CurrentPage = currentPage;
-            book.ReadingDate = readingDate;
-            book.ReadingNotes = readingNotes;
+            //want to be able to replace any data element in this edit method
+            //Need to be able to apply to correct object (id)
+            
+                Book book = context.Books.Find(editBookViewModel.Id);
+                book.Title = editBookViewModel.Title;   //breaks here
+                book.Author = editBookViewModel.Author;
+                book.TotalPage = editBookViewModel.TotalPage;
+                book.CurrentPage = editBookViewModel.CurrentPage;
+                book.ReadingDate = editBookViewModel.ReadingDate;
+                book.ReadingNotes = editBookViewModel.ReadingNotes;
 
-            return Redirect("/Books");
+                context.SaveChanges();
+                return Redirect("/Books");
+
         }
 
-        //Readbook and MoveBookMark UPDATE current page and date 
-        [HttpGet]
-        [Route("/Books/ReadBook/{bookId}")]
-        public IActionResult ReadBook(int bookId)
+            //Readbook and MoveBookMark UPDATE current page and date 
+            [HttpGet]
+            [Route("/Books/ReadBook/{bookId}")]
+        public IActionResult ReadBook(AddBookMarkViewModel addBookMarkViewModel)
         {
-            Book book = context.Books.Find(bookId);
+            Book book = context.Books.Find(addBookMarkViewModel.Id);
             ViewBag.book = book;
             ViewBag.title = $"Move Your Bookmark in {book.Title}";
- 
+
             return View();
         }
 
-        //**************** AAAAAAAAAAAAAAHHHHHHHHHHHHHH!***************
-        //TODO:  trying to update the Edit function
-        //https://docs.microsoft.com/en-us/aspnet/mvc/overview/getting-started/introduction/examining-the-edit-methods-and-edit-view
-        //https://docs.microsoft.com/en-us/aspnet/mvc/overview/getting-started/getting-started-with-ef-using-mvc/updating-related-data-with-the-entity-framework-in-an-asp-net-mvc-application
+            //https://docs.microsoft.com/en-us/aspnet/mvc/overview/getting-started/introduction/examining-the-edit-methods-and-edit-view
+            //https://docs.microsoft.com/en-us/aspnet/mvc/overview/getting-started/getting-started-with-ef-using-mvc/updating-related-data-with-the-entity-framework-in-an-asp-net-mvc-application
 
 
         [HttpPost]
         [Route("/Books/ReadingTime/")]
-        public IActionResult MoveBookMark(AddBookViewModel addBookViewModel)
+            //still need to work on datetime >>Pin<<
+            //not updating Data
+        public  IActionResult MoveBookMark(AddBookMarkViewModel addBookMarkViewModel)
         {
-            foreach (int bookId in bookIds)
+            if (ModelState.IsValid)
             {
-                Book bookMark = context.Books.Find(bookId)
-                { 
-                    CurrentPage = context.Books.Add(addBookViewModel.CurrentPage),
-                //book.ReadingDate = readingDate,
-                context.Books.ReadingNotes = readingNotes,
-            
+                Book book = new Book();
+                book.Title = addBookMarkViewModel.Title;
+                book.CurrentPage = addBookMarkViewModel.CurrentPage;
+                book.ReadingDate = addBookMarkViewModel.ReadingDate;
+                book.ReadingNotes = addBookMarkViewModel.ReadingNotes;
             }
-
+            context.SaveChanges();
             return Redirect("/Books/");
         }
 
-        //GET
-        public IActionResult Delete()
+        
+        [HttpGet]//GET
+        public IActionResult Delete(AddBookViewModel addBookViewModel)
         {
-            ViewBag.books = context.Books.ToList();
+            Book book = new Book();
             return View();
         }
 
         [HttpPost]
-        public IActionResult Delete(int[] bookIds)
+        public IActionResult DeleteBook(AddBookViewModel addBookViewModel)
         {
-            foreach(int bookId in bookIds)
+            if (ModelState.IsValid)
             {
-                Book theBook = context.Books.Find(bookId);
+                Book theBook = new Book();
                 context.Books.Remove(theBook);
             }
-            
+            context.SaveChanges();
             return Redirect("/Books/");
         }
-            
-
+        
     }
 }
+
 
 //TODO:  
     // 1. View Models?
